backend-plugin-api: make ServiceFactory backend feature
