+++ packages/backend-plugin-api/src/services/system/index.ts
@@ -15,6 +15,7 @@
  */
 
 export type {
+  BackendFeature,
   ServiceRef,
   ServiceRefConfig,
   ServiceFactory,
